cleanup and adding flip controls component
